# (#3039) - WIP DON'T MERGE - add bloomfilter to websql