--- a/lib/adapters/websql/websql-bulk-docs.js
+++ b/lib/adapters/websql/websql-bulk-docs.js
@@ -45,8 +45,8 @@ function websqlBulkDocs(req, opts, api, db, Changes, callback) {
     if (preconditionErrored) {
       return callback(preconditionErrored);
     }
-    Changes.notify(api._name);
     api._docCount = -1; // invalidate
+    Changes.notify(api._name);
     callback(null, results);
   }
 
@@ -272,6 +272,10 @@ function websqlBulkDocs(req, opts, api, db, Changes, callback) {
         return checkDone(); // skip local docs
       }
       var id = docInfo.metadata.id;
+      if (!api._docIdsFilter.test(id)) { // doesn't exist
+        api._docIdsFilter.add(id); // for races with simultaneous bulkDocs
+        return checkDone();
+      }
       tx.executeSql('SELECT json FROM ' + DOC_STORE +
       ' WHERE id = ?', [id], function (tx, result) {
         if (result.rows.length) {