--- a/lib/adapters/websql/websql.js
+++ b/lib/adapters/websql/websql.js
@@ -1,5 +1,6 @@
 'use strict';
 
+var BloomFilter = require('bloomfilter').BloomFilter;
 var utils = require('../../utils');
 var merge = require('../../merge');
 var errors = require('../../deps/errors');
@@ -61,6 +62,7 @@ function fetchAttachmentsIfNecessary(doc, opts, api, txn, cb) {
 }
 
 var cachedDatabases = {};
+var cachedDocIdsFilters = {};
 
 var openDBFunction = (typeof navigator !== 'undefined' &&
       navigator.sqlitePlugin &&
@@ -117,6 +119,11 @@ function WebSqlPouch(opts, callback) {
 
   api._docCount = -1; // cache sqlite count(*) for performance
   api._name = opts.name;
+  if (!cachedDocIdsFilters[name]) {
+    cachedDocIdsFilters[name] = new BloomFilter(8192, 16);
+  }
+  api._docIdsFilter = cachedDocIdsFilters[name];
+
 
   var db = openDB(api._name, POUCH_VERSION, api._name, size);
   if (!db) {
@@ -383,7 +390,19 @@ function WebSqlPouch(opts, callback) {
     );
   }
 
-  function onGetInstanceId() {
+  function fetchDocIds(tx) {
+    var sql = 'SELECT id FROM ' + DOC_STORE +
+      ' UNION SELECT id FROM ' + LOCAL_STORE;
+    tx.executeSql(sql, [], function (tx, res) {
+      for (var i = 0; i < res.rows.length; i++) {
+        var docId = res.rows.item(i).id;
+        api._docIdsFilter.add(docId);
+      }
+    });
+  }
+
+  function onGetInstanceId(tx) {
+    fetchDocIds(tx);
     while (idRequests.length > 0) {
       var idCallback = idRequests.pop();
       idCallback(null, instanceId);
@@ -428,7 +447,7 @@ function WebSqlPouch(opts, callback) {
             instanceId = utils.uuid();
             var initSeqArgs = [ADAPTER_VERSION, instanceId];
             tx.executeSql(initSeq, initSeqArgs, function () {
-              onGetInstanceId();
+              onGetInstanceId(tx);
             });
           });
         });
@@ -446,7 +465,7 @@ function WebSqlPouch(opts, callback) {
         var sql = 'SELECT dbid FROM ' + META_STORE;
         tx.executeSql(sql, [], function (tx, result) {
           instanceId = result.rows.item(0).dbid;
-          onGetInstanceId();
+          onGetInstanceId(tx);
         });
       };
 
@@ -548,6 +567,12 @@ function WebSqlPouch(opts, callback) {
     var doc;
     var metadata;
     var err;
+
+    // check the cache first for perf
+    if (!api._docIdsFilter.test(id)) {
+      return callback(errors.error(errors.MISSING_DOC, 'missing'));
+    }
+
     if (!opts.ctx) {
       db.readTransaction(function (txn) {
         opts.ctx = txn;
@@ -903,6 +928,12 @@ function WebSqlPouch(opts, callback) {
   };
 
   api._getLocal = function (id, callback) {
+
+    // check the cache first for perf
+    if (!api._docIdsFilter.test(id)) {
+      return callback(errors.error(errors.MISSING_DOC));
+    }
+
     db.readTransaction(function (tx) {
       var sql = 'SELECT json, rev FROM ' + LOCAL_STORE + ' WHERE id=?';
       tx.executeSql(sql, [id], function (tx, res) {
@@ -930,7 +961,12 @@ function WebSqlPouch(opts, callback) {
       newRev = doc._rev = '0-1';
     } else {
       newRev = doc._rev = '0-' + (parseInt(oldRev.split('-')[1], 10) + 1);
+      if (!api._docIdsFilter.test(doc._id)) {
+        // doc doesn't exist
+        return callback(errors.error(errors.REV_CONFLICT));
+      }
     }
+    api._docIdsFilter.add(doc._id); // prevent from using cache later
     var json = stringifyDoc(doc);
 
     var ret;