This actually causes the `temp-views` test to be a little slower (!). I think it's worth investigating why, so please don't merge this just yet.

`temp-views`, 5 iterations:

    Safari 8: 17541ms -> 18102ms
    Chrome 40: 27151ms -> 29252ms