@nolanlawson 2015-02-02
And BTW leveldb actually uses a bloom filter under the hood, and WebSQL/SQLite probably has its own caching mechanisms, but this ought to be a perf boost because we're making fewer context switches from JavaScript to native. Not sure why that's not panning out for WebSQL.