@nolanlawson 2015-02-02
Still seeing temp-views being slightly slower, but bulk-inserts is slightly faster:

bulk-inserts (500 iterations)

```
Chrome: 41067ms -> 40955ms
Safari: 29433ms -> 29305ms
```

So overall this seems like a wash, which makes me question why we need it for IndexedDB. I have no idea why a bloomfilter should be such a clear win for IndexedDB but a mixed bag for WebSQL.