@nolanlawson 2015-02-02
Hm, well 8192 may be too low to really give us gains in these tests. 1700 true negatives vs 48000 false positives.