@nolanlawson 2015-02-02
Mm, even upping the size to where we get 100% true negatives, the test time doesn't improve.