@nolanlawson 2015-02-07
see https://github.com/pouchdb/pouchdb/pull/3485#issuecomment-73372841